fixing u

+++ ft_printf/inc/ft_printf.h
@@ -85,8 +85,8 @@ int						d_handler(t_pf *cache, va_list args);
 int						u_handler(t_pf *cache, va_list args);
 int						o_handler(t_pf *cache, va_list args);
 int						x_handler(t_pf *cache, va_list args);
-int						no_case_x_handler(t_pf *cache, char *prefix, int nb,
-						char *str_nb);
+int						no_case_x_handler(t_pf *cache, char *prefix,
+						unsigned long long int nb, char *str_nb);
 int						f_handler(t_pf *cache, va_list args);
 int						p_handler(t_pf *cache, va_list args);
 int						is_spe(char x);
